Set up basic structure for the editor view - Moved src/page into src/editor/page - Added basic components for the editor, (the constituent markdown and html editor) - Added CSS modules
