+++ package.json
@@ -7,6 +7,8 @@
     "@types/react": "^16.7.3",
     "@types/react-dom": "^16.0.9",
     "@types/react-redux": "^6.0.9",
+    "@uifabric/fluent-theme": "^0.13.0",
+    "office-ui-fabric-react": "^6.118.2",
     "react": "^16.6.1",
     "react-dom": "^16.6.1",
     "react-redux": "^5.1.1",
