+++ src/App.tsx
@@ -1,27 +1,10 @@
 import React, { Component } from "react";
 import "./App.css";
-import logo from "./logo.svg";
+import { EditorComponent } from "./editor/editor";
 
 class App extends Component {
   public render(): JSX.Element {
-    return (
-      <div className="App">
-        <header className="App-header">
-          <img src={logo} className="App-logo" alt="logo" />
-          <p>
-            Edit <code>src/App.js</code> and save to reload.
-          </p>
-          <a
-            className="App-link"
-            href="https://reactjs.org"
-            target="_blank"
-            rel="noopener noreferrer"
-          >
-            Learn React
-          </a>
-        </header>
-      </div>
-    );
+    return <EditorComponent />;
   }
 }
