+++ src/editor/htmlEditor/htmlEditor.module.css
@@ -1,0 +1,6 @@
+.htmlEditor {
+  float: right;
+  width: 50%;
+  height: 100vh;
+  background-color: pink;
+}
