+++ src/editor/markdownEditor/markdownEditor.module.css
@@ -1,0 +1,6 @@
+.markdownEditor {
+  float: left;
+  width: 50%;
+  height: 100vh;
+  background-color: aquamarine;
+}
