+++ src/editor/page/tree/delete.test.ts
