+++ src/editor/page/tree/delete.ts
